Fix critical color display bugs and improve pan sensitivity
## Color System Fixes
- **Fixed Highlighter Dark Grey Issue**: Corrected color conversion from Color.value.toLong() to Color.toArgb().toUInt().toLong()
- **Fixed Pen Yellow Display Issue**: Same color conversion fix ensures all drawing tools display correct selected colors
- **Improved Color Retrieval**: Changed Color(stroke.color.toULong()) to Color(stroke.color.toUInt().toInt()) for accurate reconstruction

## Pan/Zoom Improvements
- **Scale-Aware Pan Sensitivity**: Pan now adjusts to current zoom level for natural finger tracking
  - When zoomed out: Slower, more precise pan movement
  - When zoomed in: Faster pan that matches finger movement
- **Applied to Both Modes**: Fixed in both view mode and drawing mode pan gestures

## Technical Details
- **Color Pipeline**: Compose Color → ARGB Int → UInt → Long → UInt → ARGB Int → Compose Color
- **Pan Formula**: offset += pan * currentScale (was: offset += pan)
- **Type Safety**: Added proper UInt/Int conversions to prevent compilation errors

## Files Updated
- AnnotationCanvas.kt: Stroke creation and display color fixes
- AnnotatableFileViewer.kt: Pan sensitivity and overlay color fixes
- AnnotationToolbar.kt: Color circles in stroke list
- FileViewerViewModel.kt: Text annotation color storage
- Annotation.kt: Default color value correction

Resolves color corruption issues where strokes appeared in wrong colors (grey highlighter, yellow pen) and improves zoom/pan user experience with scale-responsive sensitivity.

+++ app/src/main/java/com/troubashare/ui/components/AnnotationToolbar.kt
@@ -440,7 +440,7 @@ fun AnnotationListItem(
                             .size(16.dp)
                             .background(
                                 try {
-                                    androidx.compose.ui.graphics.Color(stroke.color.toULong())
+                                    androidx.compose.ui.graphics.Color(stroke.color.toUInt().toInt())
                                 } catch (e: Exception) {
                                     androidx.compose.ui.graphics.Color.Black
                                 },
@@ -530,7 +530,7 @@ fun SelectableAnnotationListItem(
                                 .size(16.dp)
                                 .background(
                                     try {
-                                        androidx.compose.ui.graphics.Color(stroke.color.toULong())
+                                        androidx.compose.ui.graphics.Color(stroke.color.toUInt().toInt())
                                     } catch (e: Exception) {
                                         androidx.compose.ui.graphics.Color.Black
                                     },
